디테일 페이지 1차 기능구현완료
