+++ src/domains/course/hooks/useCourseDetail.tsx
@@ -1,36 +1,36 @@
-import { useEffect, useState } from 'react'
-import { getCourse } from '../services/courseService'
-import { CourseDetailResponse } from '../types/types'
+import { useEffect, useState } from 'react';
+import { getCourse } from '../services/courseService';
+import { CourseDetailResponse } from '../types/course';
 
 export function useCourseDetail(courseId: string) {
   const [courseData, setCourseData] = useState<CourseDetailResponse>({
     course: null,
     sections: [],
     lectures: {},
-  })
+  });
 
-  const [loading, setLoading] = useState<boolean>(true)
+  const [loading, setLoading] = useState<boolean>(true);
 
   useEffect(() => {
-    if (!courseId) return
+    if (!courseId) return;
 
     const fetchCourseDetail = async () => {
       try {
-        setLoading(true)
-        const data = await getCourse(courseId)
-        setCourseData(data)
+        setLoading(true);
+        const data = await getCourse(courseId);
+        setCourseData(data);
       } catch (err) {
-        console.error('강좌 상세 정보 불러오기 실패:', err)
+        console.error('강좌 상세 정보 불러오기 실패:', err);
       } finally {
-        setLoading(false)
+        setLoading(false);
       }
-    }
+    };
 
-    fetchCourseDetail()
-  }, [courseId])
+    fetchCourseDetail();
+  }, [courseId]);
 
   return {
     ...courseData,
     loading,
-  }
+  };
 }
