+++ src/domains/course/utils/formatDuration.ts
@@ -1,13 +1,10 @@
-// src/domains/course/utils/formatDuration.ts
+export const formatDuration = (totalMinutes: number): string => {
+  if (!totalMinutes || totalMinutes <= 0) return '0분';
 
-// 총 분(minute)을 "1시간 20분" 같은 문자열로 변환
-export const formatDuration = (minutes: number): string => {
-  if (!minutes || minutes <= 0) return '0분'
+  const hours = Math.floor(totalMinutes / 60);
+  const minutes = totalMinutes % 60;
 
-  const hours = Math.floor(minutes / 60)
-  const mins = minutes % 60
-
-  if (hours === 0) return `${mins}분`
-  if (mins === 0) return `${hours}시간`
-  return `${hours}시간 ${mins}분`
-}
+  if (hours === 0) return `${minutes}분`;
+  if (minutes === 0) return `${hours}시간`;
+  return `${hours}시간 ${minutes}분`;
+};
